+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 ## 1.0.4 (WIP)
 
 * Don't create a `db` instance when the MongoDatasource fails to connect
+* Add `Document#$attributes` property, where non-database values can be stored in
+* Add `Document#hasChanged(name)` method, to check if values have changed
+* Add `Document#resetFields()` to reset a document to its initial state
 
 ## 1.0.3 (2018-07-12)